flask_projects/day-62-starting-files-coffee-and-wifi/requirements.txt
python-dotenv
Bootstrap_Flask==2.2.0
Flask==2.3.2
WTForms==3.0.1
Flask_WTF==1.2.1
Werkzeug==3.0.0